# solution/docs/swagger.yaml
components:
  schemas:
    dto.AdUser:
      properties:
        ad_id:
          type: string
        ad_text:
          type: string
        ad_title:
          type: string
        advertiser_id:
          type: string
      type: object
...
    dto.QueryClient:
      properties:
        client_id:
          form: client_id
          type: string
      required:
      - client_id
      type: object
...
    models.Advertiser:
      properties:
        advertiser_id:
          type: string
        name:
          type: string
      required:
      - advertiser_id
      - name
      type: object
    models.Campaign:
      properties:
        ad_text:
          type: string
        ad_title:
          type: string
        advertiser_id:
          type: string
        campaign_id:
          type: string
        clicks_limit:
          type: integer
        cost_per_click:
          type: number
        cost_per_impression:
          type: number
        end_date:
          type: integer
        image:
          type: string
        impressions_limit:
          type: integer
        moderated:
          type: boolean
        start_date:
          type: integer
        targeting:
          $ref: '#/components/schemas/models.Targeting'
      required:
      - ad_text
      - ad_title
      - clicks_limit
      - cost_per_click
      - cost_per_impression
      - end_date
      - impressions_limit
      - start_date
      type: object
    models.CampaignUpdate:
      properties:
        ad_text:
          type: string
        ad_title:
          type: string
        advertiser_id:
          type: string
        campaign_id:
          type: string
        clicks_limit:
          type: integer
        cost_per_click:
          type: number
        cost_per_impression:
          type: number
        impressions_limit:
          type: integer
        targeting:
          $ref: '#/components/schemas/models.Targeting'
      required:
      - ad_text
      - ad_title
      - clicks_limit
      - impressions_limit
      type: object
...
    models.Targeting:
      properties:
        age_from:
          type: integer
        age_to:
          type: integer
        gender:
          enum:
          - MALE
          - FEMALE
          - ALL
          type: string
        location:
          type: string
      type: object
...
info:
  description: PROOOOOOOOOD
  title: Prod-2025 AD API
  version: "1.0"
openapi: 3.1.0
paths:
  /ads:
    get:
      description: Get ads by client ID
      parameters:
      - description: Client ID
        in: query
        name: client_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.AdUser'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get ads
      tags:
      - ads
  /ads/{id}/click:
    post:
      description: Register a click on an ad by ID
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.QueryClient'
        description: Client ID
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Click an ad
      tags:
      - ads
  /advertisers/{id}:
    get:
      description: Get an advertiser by ID
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Advertiser'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get an advertiser
      tags:
      - advertisers
  /advertisers/{id}/campaigns:
    get:
      description: Get campaigns by advertiser ID with pagination
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: Page size
        in: query
        name: size
        required: true
        schema:
          type: integer
      - description: Page number
        in: query
        name: page
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Campaign'
                type: array
          description: OK
          headers:
            X-Total-Count:
              description: Total number of campaigns
              schema:
                type: integer
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: List campaigns
      tags:
      - campaigns
    post:
      description: Create a new campaign for an advertiser
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.Campaign'
        description: Campaign data
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Campaign'
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
      summary: Create a campaign
      tags:
      - campaigns
  /advertisers/{id}/campaigns/{camp_id}:
    delete:
      description: Delete a campaign by advertiser ID and campaign ID
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: Campaign ID
        in: path
        name: camp_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Delete a campaign
      tags:
      - campaigns
    get:
      description: Get a campaign by advertiser ID and campaign ID
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: Campaign ID
        in: path
        name: camp_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Campaign'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get a campaign
      tags:
      - campaigns
    put:
      description: Update a campaign by advertiser ID and campaign ID
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: Campaign ID
        in: path
        name: camp_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.CampaignUpdate'
        description: Campaign update data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Campaign'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Conflict
      summary: Update a campaign
      tags:
      - campaigns
...
  /advertisers/bulk:
    post:
      description: Bulk create advertisers
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/models.Advertiser'
              type: array
        description: List of advertisers
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Advertiser'
                type: array
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
      summary: Bulk create advertisers
      tags:
      - advertisers
...
servers:
- url: http://localhost:8080/